cleanup aui/gui/wv code add new type wnd_colidx for column index into line data array correct prototypes

+++ bld/aui/c/wndgadgt.c
@@ -34,16 +34,17 @@
 
 void WndGetGadgetSize( wnd_gadget_type type, gui_coord *size )
 {
-    GUIGetHotSpotSize( type+1, size );
+    GUIGetHotSpotSize( type + 1, size );
 }
 
-void WndSetGadgetLine( a_window wnd, wnd_line_piece *line, wnd_gadget_type type, unsigned length )
+void WndSetGadgetLine( a_window wnd, wnd_line_piece *line, wnd_gadget_type type, size_t length )
 {
     static char text[2] = " ";
 
     line->hint = WndGadgetHint[type];
-    if( WndSwitchOn( wnd, WSW_ALTERNATE_BIT ) ) type += WndGadgetSecondary;
-    text[0] = type+1;
+    if( WndSwitchOn( wnd, WSW_ALTERNATE_BIT ) )
+        type += WndGadgetSecondary;
+    text[0] = type + 1;
     line->text = text;
     line->bitmap = true;
     line->length = length;
